refactor(e2e): modularize MCP server types and utilities
Split monolithic types.ts into focused modules:
- types/: 9 focused type modules (responses, errors, seeding, etc.)
- utils/: 5 utility modules (response, redaction, targets, time)
- tools/helpers.ts: shared tool handler utilities
- tools/registry.ts: handler registry replacing switch statement

Key improvements:
- Eliminated ~277 lines of duplication
- Fixed circular dependency between schemas.ts and types.ts
- Server.ts now uses registry pattern instead of 22-case switch
- Added unit tests for new utilities (registry, response, targets, time, redaction)
- Removed obsolete test files (example-usage, test-*.ts)

See SPEC-06.md for detailed specification.
