+++ test/e2e/playwright/llm-workflow/mcp-server/types/errors.ts
@@ -1,0 +1,29 @@
+export const ErrorCodes = {
+  MM_BUILD_FAILED: 'MM_BUILD_FAILED',
+  MM_DEPENDENCIES_MISSING: 'MM_DEPENDENCIES_MISSING',
+
+  MM_SESSION_ALREADY_RUNNING: 'MM_SESSION_ALREADY_RUNNING',
+  MM_NO_ACTIVE_SESSION: 'MM_NO_ACTIVE_SESSION',
+  MM_LAUNCH_FAILED: 'MM_LAUNCH_FAILED',
+
+  MM_INVALID_CONFIG: 'MM_INVALID_CONFIG',
+  MM_INVALID_INPUT: 'MM_INVALID_INPUT',
+  MM_PORT_IN_USE: 'MM_PORT_IN_USE',
+
+  MM_NAVIGATION_FAILED: 'MM_NAVIGATION_FAILED',
+  MM_NOTIFICATION_TIMEOUT: 'MM_NOTIFICATION_TIMEOUT',
+
+  MM_TARGET_NOT_FOUND: 'MM_TARGET_NOT_FOUND',
+  MM_CLICK_FAILED: 'MM_CLICK_FAILED',
+  MM_TYPE_FAILED: 'MM_TYPE_FAILED',
+  MM_WAIT_TIMEOUT: 'MM_WAIT_TIMEOUT',
+
+  MM_SCREENSHOT_FAILED: 'MM_SCREENSHOT_FAILED',
+
+  MM_KNOWLEDGE_ERROR: 'MM_KNOWLEDGE_ERROR',
+
+  MM_SEED_FAILED: 'MM_SEED_FAILED',
+  MM_CONTRACT_NOT_FOUND: 'MM_CONTRACT_NOT_FOUND',
+} as const;
+
+export type ErrorCode = (typeof ErrorCodes)[keyof typeof ErrorCodes];
