+++ test/e2e/playwright/llm-workflow/mcp-server/types/session.ts
@@ -1,0 +1,10 @@
+export type SessionState = {
+  sessionId: string;
+  extensionId: string;
+  startedAt: string;
+  ports: {
+    anvil: number;
+    fixtureServer: number;
+  };
+  stateMode: 'default' | 'onboarding' | 'custom';
+};
