+++ test/e2e/playwright/llm-workflow/mcp-server/utils/redaction.ts
@@ -1,0 +1,13 @@
+export const SENSITIVE_FIELD_PATTERNS = [
+  /password/iu,
+  /seed/iu,
+  /srp/iu,
+  /phrase/iu,
+  /mnemonic/iu,
+  /private.*key/iu,
+  /secret/iu,
+] as const;
+
+export function isSensitiveField(fieldName: string): boolean {
+  return SENSITIVE_FIELD_PATTERNS.some((pattern) => pattern.test(fieldName));
+}
